refactor: Refactoring imports, functions and add documents
