+++ .env.example
@@ -1,7 +1,12 @@
+# Project
+ENVIRONMENT='development'
 NODE_ENV_MESSAGE=''
-MONGODB_URI=''
 PORT=
 
+# Database
+MONGODB_URI=''
+
+#Auth
 JWT_ISSUER=''
 JWT_SECRET_PUBLIC=``
 JWT_SECRET_PRIVATE=``
