+++ src/api/auth/jwt/jwt.strategy.ts
@@ -1,15 +1,16 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable } from '@nestjs/common';
 import { ExtractJwt, Strategy } from 'passport-jwt';
-import { PassportStrategy } from "@nestjs/passport";
-import { AuthService } from "../auth.service";
+import { PassportStrategy } from '@nestjs/passport';
+import { AuthService } from '../auth.service';
+import { authEnv } from 'src/infra/env';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
   constructor(private authService: AuthService) {
     super({
       jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-      ignoreExpiration: false, 
-      secretOrKey: process.env.JWT_SECRET_PUBLIC,
+      ignoreExpiration: false,
+      secretOrKey: authEnv.jwt.secretPublic,
       algorithms: ['RS256'],
     });
   }
