+++ src/app.controller.ts
@@ -1,5 +1,6 @@
 import { Controller, Get } from '@nestjs/common';
-import { AppService } from './app.service';
+
+import { AppService } from 'src/app.service';
 
 @Controller()
 export class AppController {
