+++ src/app.module.ts
@@ -1,18 +1,20 @@
-import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ConfigModule } from '@nestjs/config';
-import { AuthModule, UsersModule } from './api/module-exporter';
+import { Module } from '@nestjs/common';
+
+import { AuthModule, UsersModule } from 'src/api/module-exporter';
+import { AppController } from 'src/app.controller';
+import { AppService } from 'src/app.service';
+import { mongoEnv } from './infra/env';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
     }),
-    MongooseModule.forRoot(String(process.env.MONGODB_URI)),
+    MongooseModule.forRoot(mongoEnv.uri),
     AuthModule,
-    UsersModule
+    UsersModule,
   ],
   controllers: [AppController],
   providers: [AppService],
