+++ src/utils/functions/index.ts
@@ -1,21 +1,12 @@
-import * as bcrypt from 'bcrypt';
-import { ERole } from 'src/models/roles';
-
-export const hashPassword = async (password: string): Promise<string> => {
-  const salt = await bcrypt.genSalt();
-  const hash = await bcrypt.hash(password, salt);
-
-  return hash;
-};
-
-export const comparePassword = async (
-  password: string,
-  hash: string,
-): Promise<boolean> => {
-  return await bcrypt.compare(password, hash);
-};
-
-export const generateCode = () => {
+/**
+ * Generates a random alphanumeric code consisting of 6 characters.
+ *
+ * The generated code is composed of uppercase letters (A-Z) and digits (0-9),
+ * ensuring a total of 36 possible characters for each position in the string.
+ *
+ * @returns {string} A randomly generated 6-character alphanumeric string.
+ */
+export const generateCode = (): string => {
   const character = '01234ABCDEFGHIJKLMNOPQRSTUVWXYZ56789';
   let generateString = '';
 
@@ -28,24 +19,18 @@ export const generateCode = () => {
   return generateString;
 };
 
-export const timeout = (ms: number) => {
+/**
+ * Delays the execution of code for a specified amount of time.
+ *
+ * @param ms - The number of milliseconds to wait before resolving the promise.
+ * @returns A promise that resolves after the specified delay.
+ *
+ * @example
+ * ```typescript
+ * await timeout(1000); // Waits for 1 second before proceeding
+ * console.log('1 second has passed');
+ * ```
+ */
+export const timeout = (ms: number): Promise<unknown> => {
   return new Promise((resolve) => setTimeout(resolve, ms));
 };
-
-export const parseJwt = (token: string) => {
-  const jwt = token.replace('Bearer ', '');
-  return JSON.parse(String(Buffer.from(jwt.split('.')[1], 'base64')));
-};
-
-export const isSameUser = (id: string, auth: string) => {
-  const { sub } = parseJwt(auth);
-  if (id !== sub) return false;
-  return true;
-};
-
-export const isUserAdmin = (auth: string) => {
-    const { roles } = parseJwt(auth);
-    const isAllowed = roles.includes(ERole.admin);
-  
-    return isAllowed;
-  }
