+++ src/utils/index.ts
@@ -1,9 +1,3 @@
-export { textResponse, dataResponse } from './response';
-export {
-    hashPassword,
-    comparePassword,
-    generateCode,
-    timeout,
-    isSameUser,
-    isUserAdmin
-} from './functions';
+export * from './response';
+export * from './auth';
+export * from './functions';
